cleaning up repo and fixing errors
